refactor: WinningResult클래스의 함수 구조 추가 변경

+++ src/main/java/model/winning/WinningResult.java
@@ -17,7 +17,7 @@ public class WinningResult {
 
     public Double getRateOfReturn(final Money money) {
         int totalInsertMoney = money.getMoney();
-        double totalReturn = calculateTotalReturn();
+        int totalReturn = calculateTotalReturn();
         return totalReturn / (double) totalInsertMoney;
     }
